Add split-resizing shortcut for smaller monitor
I got used to this vertical-split setup (code on one side, tests on the
other) while coding on a Thunderbolt Display, but when I moved my Vim
windows to a smaller Cinema Display, I found out I couldn't see both
files at full width anymore.

(Where "full width" means that each file has its gitgutter, line numbers
and all 79 characters of a typical line visible without needing to
scroll to the right. This is mostly a problem because I choose to set my
text size so high.)

This commit introduces a \r shortcut that resizes the current file's
split so I can see it at full width, although that means the file on
the other size of the split will be a bit less visible.